refactor: Service Exception 수정
